+++ src/helpers.rs
@@ -92,7 +92,7 @@ impl<S> Handler<S> for NormalizePath {
             // merge slashes
             let p = self.re_merge.replace_all(req.path(), "/");
             if p.len() != req.path().len() {
-                if req.resource().has_prefixed_route(p.as_ref()) {
+                if req.resource().has_prefixed_resource(p.as_ref()) {
                     let p = if !query.is_empty() {
                         p + "?" + query
                     } else {
@@ -105,7 +105,7 @@ impl<S> Handler<S> for NormalizePath {
                 // merge slashes and append trailing slash
                 if self.append && !p.ends_with('/') {
                     let p = p.as_ref().to_owned() + "/";
-                    if req.resource().has_prefixed_route(&p) {
+                    if req.resource().has_prefixed_resource(&p) {
                         let p = if !query.is_empty() {
                             p + "?" + query
                         } else {
@@ -120,7 +120,7 @@ impl<S> Handler<S> for NormalizePath {
                 // try to remove trailing slash
                 if p.ends_with('/') {
                     let p = p.as_ref().trim_right_matches('/');
-                    if req.resource().has_prefixed_route(p) {
+                    if req.resource().has_prefixed_resource(p) {
                         let mut req = HttpResponse::build(self.redirect);
                         return if !query.is_empty() {
                             req.header(
@@ -135,7 +135,7 @@ impl<S> Handler<S> for NormalizePath {
             } else if p.ends_with('/') {
                 // try to remove trailing slash
                 let p = p.as_ref().trim_right_matches('/');
-                if req.resource().has_prefixed_route(p) {
+                if req.resource().has_prefixed_resource(p) {
                     let mut req = HttpResponse::build(self.redirect);
                     return if !query.is_empty() {
                         req.header(
@@ -151,7 +151,7 @@ impl<S> Handler<S> for NormalizePath {
         // append trailing slash
         if self.append && !req.path().ends_with('/') {
             let p = req.path().to_owned() + "/";
-            if req.resource().has_prefixed_route(&p) {
+            if req.resource().has_prefixed_resource(&p) {
                 let p = if !query.is_empty() {
                     p + "?" + query
                 } else {